+++ src/BenchmarkDotNet.Diagnostics.Windows/PmcDiagnoser.cs
@@ -4,11 +4,9 @@ using System.Linq;
 using BenchmarkDotNet.Columns;
 using BenchmarkDotNet.Diagnosers;
 using BenchmarkDotNet.Engines;
-using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Loggers;
 using BenchmarkDotNet.Reports;
 using BenchmarkDotNet.Running;
-using BenchmarkDotNet.Toolchains.InProcess;
 using BenchmarkDotNet.Validators;
 using Microsoft.Diagnostics.Tracing.Parsers;
 using Microsoft.Diagnostics.Tracing.Parsers.Kernel;
@@ -18,24 +16,6 @@ namespace BenchmarkDotNet.Diagnostics.Windows
 {
     public class PmcDiagnoser : EtwDiagnoser<PmcStats>, IHardwareCountersDiagnoser
     {
-        internal static readonly Dictionary<HardwareCounter, string> EtwTranslations
-            = new Dictionary<HardwareCounter, string>
-            {
-                { HardwareCounter.Timer, "Timer" },
-                { HardwareCounter.TotalIssues, "TotalIssues" },
-                { HardwareCounter.BranchInstructions, "BranchInstructions" },
-                { HardwareCounter.CacheMisses, "CacheMisses" },
-                { HardwareCounter.BranchMispredictions, "BranchMispredictions" },
-                { HardwareCounter.TotalCycles, "TotalCycles" },
-                { HardwareCounter.UnhaltedCoreCycles, "UnhaltedCoreCycles" },
-                { HardwareCounter.InstructionRetired, "InstructionRetired" },
-                { HardwareCounter.UnhaltedReferenceCycles, "UnhaltedReferenceCycles" },
-                { HardwareCounter.LlcReference, "LLCReference" },
-                { HardwareCounter.LlcMisses, "LLCMisses" },
-                { HardwareCounter.BranchInstructionRetired, "BranchInstructionRetired" },
-                { HardwareCounter.BranchMispredictsRetired, "BranchMispredictsRetired" }
-            };
-
         private readonly Dictionary<BenchmarkCase, PmcStats> results = new Dictionary<BenchmarkCase, PmcStats>();
 
         // ReSharper disable once EmptyConstructor parameterless ctor is mandatory for DiagnosersLoader.CreateDiagnoser
@@ -72,53 +52,15 @@ namespace BenchmarkDotNet.Diagnostics.Windows
         public void DisplayResults(ILogger logger) { }
 
         public IEnumerable<ValidationError> Validate(ValidationParameters validationParameters)
-        {
-            if (!validationParameters.Config.GetHardwareCounters().Any())
-            {
-                yield return new ValidationError(true, "No Hardware Counters defined, probably a bug");
-                yield break;
-            }
-
-            if (TraceEventSession.IsElevated() != true)
-                yield return new ValidationError(true, "Must be elevated (Admin) to use Hardware Counters to use ETW Kernel Session.");
-
-            var availableCpuCounters = TraceEventProfileSources.GetInfo();
-
-            foreach (var hardwareCounter in validationParameters.Config.GetHardwareCounters())
-            {
-                if (!EtwTranslations.TryGetValue(hardwareCounter, out var counterName))
-                    yield return new ValidationError(true, $"Counter {hardwareCounter} not recognized. Please make sure that you are using counter available on your machine. You can get the list of available counters by running `tracelog.exe -profilesources Help`");
-
-                if (!availableCpuCounters.ContainsKey(counterName))
-                    yield return new ValidationError(true, $"The counter {counterName} is not available. Please make sure you are Windows 8+ without Hyper-V");
-            }
-
-            foreach (var benchmark in validationParameters.Benchmarks)
-            {
-                if (benchmark.Job.Infrastructure.HasValue(InfrastructureMode.ToolchainCharacteristic)
-                    && benchmark.Job.Infrastructure.Toolchain is InProcessToolchain)
-                {
-                    yield return new ValidationError(true, "Hardware Counters are not supported for InProcessToolchain.", benchmark);
-                }
-            }
-        }
+            => HardwareCounters.Validate(validationParameters, mandatory: true);
 
         protected override PmcStats GetInitializedStats(DiagnoserActionParameters parameters)
         {
-            var stats = new PmcStats(parameters.Config.GetHardwareCounters().ToArray(), FromCounter);
+            var stats = new PmcStats(
+                parameters.Config.GetHardwareCounters().ToArray(), 
+                counter => HardwareCounters.FromCounter(counter, info => info.MinInterval )); // for this diagnoser we want the smallest interval to have best possible precision
 
-            var counters = stats.Counters.Values;
-
-            try
-            {
-                TraceEventProfileSources.Set( // it's a must have to get the events enabled!!
-                    counters.Select(counter => counter.ProfileSourceId).ToArray(),
-                    counters.Select(counter => counter.Interval).ToArray());
-            }
-            catch (System.Runtime.InteropServices.COMException ex) when (ex.Message.StartsWith("The WMI data block or event notification has already been enabled"))
-            {
-                // previous run was interrupted by ctrl+c and never stopped
-            }
+            HardwareCounters.Enable(stats.Counters.Values);
 
             return stats;
         }
@@ -139,14 +81,6 @@ namespace BenchmarkDotNet.Diagnostics.Windows
         {
             if (StatsPerProcess.TryGetValue(obj.ProcessID, out var stats))
                 stats.Handle(obj.ProfileSource, obj.InstructionPointer);
-        }
-
-        private static PreciseMachineCounter FromCounter(HardwareCounter counter)
-        {
-            var profileSource = TraceEventProfileSources.GetInfo()[EtwTranslations[counter]]; // it can't fail, diagnoser validates that first
-
-            return new PreciseMachineCounter(profileSource.ID, profileSource.Name, counter,
-                profileSource.MinInterval); // we want the smallest interval to have best possible precision
         }
 
         public IColumnProvider GetColumnProvider()
@@ -255,6 +189,5 @@ namespace BenchmarkDotNet.Diagnostics.Windows
                     ? (stats.Counters[HardwareCounter.InstructionRetired].Count / (double)stats.Counters[HardwareCounter.TotalCycles].Count).ToString("N2")
                     : "-";
         }
-
     }
 }
